fix(synapse-bridge): leer RECID como Long
Los RecId de D365 son bigint y todos superan el rango de int (p. ej. los
19.161 de inventtable están entre 5.637.161.826 y 5.638.272.576), así que
rs.getInt("RECID") fallaba en /products/items. EcomProductDTO.id pasa a Long.

El contrato no cambia: el JSON sigue siendo un número y core-orchestrator lo
decodifica en domain.Product.Id (int de 64 bits en linux/amd64); además no lo
usa, identifica los productos por Code.

+++ services/synapse-bridge/src/main/java/com/vegusa/ecommerce/dto/EcomProductDTO.java
@@ -2,7 +2,7 @@ package com.vegusa.ecommerce.dto;
 
 public record EcomProductDTO(
         SourceSystem source,
-        Integer id, // RECID
+        Long id, // RECID (bigint en D365: los RecId superan el rango de int)
         String code, // Articulo
         String description, // Descripción
         String partNumber, // Num.Parte
